Attempting to optimise gains of PID controller as precisions of an active inference process.
The maths suggests that the gains should encode the observation noise of the process (sigma^2_w), but if motor control is allowed during the estimation of the gain, the action will push the expected mean of the observations towards the prior (desire, set point) rather than the true mean, thus providing a wrong estimation of the variance.

At the moment I'm testing the idea of having the precisions estimated on a freely moving agent and then having such agent moving after a good estimate is available. The estimation of precisions somehow works but it is not very robust and it is quite sensitive to the parametrisation of the minimisation procedure (learning rates and damping for this specific process). I can get it to work but there's a general tendency for the agent to overestimate the precisions, which might help later on for the control aspect so this is not a problem as of now. Action works quite well with the estimated precision BUT after the training this agent is in a state (position and velocity) that is not controllable (the car is going too fast?) for this setup to work (action/control is not bounded) so this agent "overreacts" due to the very poor state it finds itself in after the estimation of precisions is frozen.