The optimisation of the gains as precisions seems to be definitely working. However, it is dependent on the optimisation of the log-precision on priors: if they are too tight, action won't be able to counteract the effects of some basic level of noise which leads to an over-estimation of the covariance matrix of the sensations since it needs to account for all the fluctuations. This implied a decrease of the log-precisions on sensory predictions errors/gains of the PID controller, bringing the controller to a different (higher-variance) equilibrium.
It looks like this method is simply shifting the load to the optimisation of log-precisions on priors, not making the problem any easier. If, in fact we just try to optimise the latter together with all the system, a stable solution will be somewhere between a good inference and a good control (not good for engineers then).